Enable ECDSA ciphers when running with FIPS cryptography
In FIPS mode, certain cryptographic algorithms are typically disabled
this includes the algorithm `SHA1withECDSA` since `SHA1` is no longer
recommended for general use.

The `JSSE`provider, responsible for managing TLS cipher specifications,
currently disables all `*ECDSA*` ciphers due to the absence of the
`SHA1withECDSA` algorithm on the system when running in fips mode.
This behavior is inaccurate, since other ECDSA-related
transformations, such as `SHA512withECDSA` or `SHA384withECDSA`, are
available in a FIPS environment and can be used for TLS connections.

With this update, `*ECDSA*` ciphers can now be enabled in FIPS mode
without the existing check that `SHA1withECDSA` is present. We will now
check for the presence of `SHA512withECDSA` instead which is available
in both the NSS and OpenJCEPlus FIPS solutions being used in Semeru.